HistoryInfo,VisitHistoy modify
