+++ src/components/HistoryInfo.jsx
@@ -1,6 +1,5 @@
-import React, { useState } from 'react';
+
 import styled from 'styled-components';
-import { Link, useNavigate } from 'react-router-dom';
 
 const WrapHistory = styled.div`
   display: flex;
@@ -24,30 +23,5 @@ const TitleText = styled.div`
   font-weight: bold;
   left: 20px;
 `;
-export default function HistoryInfo() {
-  const navigate = useNavigate();
-  const [history, setHistory] = useState(JSON.parse(sessionStorage.getItem('history')));
-  const handleClick = (title) => {
-    navigate(`/AuctionDetail/${title}`);
-  }; //주소 이동
 
-  return (
-    <>
-      <WrapHistory>
-        {history.map((id, index) => (
-          <HistoryBox key={index} onClick={() => handleClick(id)}>
-            <TitleText>{id}</TitleText>
-          </HistoryBox>
-        ))}
-        <button
-          onClick={() => {
-            sessionStorage.setItem('history', JSON.stringify([]));
-            setHistory([]);
-          }}
-        >
-          Clear History
-        </button>
-      </WrapHistory>
-    </>
-  );
-}
+export { WrapHistory, HistoryBox, TitleText };
